Replace reference to Bitnami OpenLDAP image
Bitname decided to limit distribution of such an image to a commercial
subscription. Needless to say, this project isn't going to pay for such
a subscription.

There is an alternative image packaged by CleanStart, but it lacks the
amenities used in this codebase for setting up the LDAP server. I'd
also like to avoid migrating to e.g. LLDAP because that's a too big
endeavor for the purposes of getting CI working again.

Therefore, let's switch to the `bitnamilegacy` organization for now, as
having a very up to date OpenLDAP server is not critical for development
and running tests.